Add social links to Author page
